Concatenate multiple .covClasss classes

Concatenate multiple .covClasss classes

Usage

# S4 method for class '.covClass'
c(x, ...)

Arguments

x

A .covClasss class with covariate information generated in set_cov

...

Other .covClasss classes with covariate information generated in set_cov

Value

A vector of .covClasss classes

Examples

# combine two sets of covariates
covset1 = set_cov(n_cat = 2, n_cont = 0, mu_int = 0, mu_ext = 0, var = 1)
#> User provides one mean of covariate in internal trials. This value is used for all covariates
#> User provides one mean of covariate in external trials. This value is used for all covariates
#> User provides one number for variance. This variance is used for all covariates
#> Covariance for covariate (cov) is not recognized. Default value 0 is used for all covariates
#> Probability of binary covariate in the internal trial is not recognized or correctly specified. Default value 0.5 is used for all binary covariates
#> Probability of binary covariate in the external trial is not recognized or correctly specified. mu_int is used.
covset2 = set_cov(n_cat = 0, n_cont = 1, mu_int = 62, mu_ext = 65, var = 11)
cov_list = c(covset1, covset2)