Match

Match

Usage

match_cov(dt, match)

Arguments

dt

a list of matrix

match

A vector of covariates name to match on

Value

a list of matrix containing matched cohort information

Examples

# match internal and external trial data using different covariates
smp = set_n(ssC = 140, ssE = 275, ssExt = 100)
covset1 = set_cov(n_cat = 2, n_cont = 0, mu_int = 0, mu_ext = 0, var = 1)
#> User provides one mean of covariate in internal trials. This value is used for all covariates
#> User provides one mean of covariate in external trials. This value is used for all covariates
#> User provides one number for variance. This variance is used for all covariates
#> Covariance for covariate (cov) is not recognized. Default value 0 is used for all covariates
#> Probability of binary covariate in the internal trial is not recognized or correctly specified. Default value 0.5 is used for all binary covariates
#> Probability of binary covariate in the external trial is not recognized or correctly specified. mu_int is used.
covset2 = set_cov(n_cat = 0, n_cont = 1, mu_int = 62, mu_ext = 65, var = 11)
cObj = c(covset1, covset2)
sample_cov <-
  simu_cov(ssObj = smp, covObj = cObj, HR = 1, driftHR = 1.2, nsim = 2)
#> The sample size for internal control, internal treatment, and external control arms are 140, 275, and 100, respectively.
#> Error in paste0("Set seed to ", .Random.seed[1]): object '.Random.seed' not found

# match on covariates 1 and 2
match_cov(dt = sample_cov, match = c("cov1", "cov2"))
#> Error in is.data.frame(x): object 'sample_cov' not found

# match on all 3 covariates
match_cov(dt = sample_cov, match = c("cov1", "cov2", "cov3"))
#> Error in is.data.frame(x): object 'sample_cov' not found