Specify parameters for enrollment time, drop-out pattern and analysis start time

This function allows user to specify the enrollment and drop-out rate, and the type of clinical cut-off Date. Both enrollment times and drop-out times follow piece-wise exponential distribution.

Usage

set_clin(gamma, e_itv, CCOD, CCOD_t, etaC, etaE, d_itv)

Arguments

gamma

A vector of rate of enrollment per unit of time

e_itv

A vector of duration of time periods for recruitment with rates specified in gamma. Note that the length of e_itv should be same length as gamma or 1 less.

CCOD

Type of analysis start time. Analysis starts at CCOD_t months after the first or last patient's enrollment if CCOD = "fixed-first" or CCOD = "fixed-last" respectively. Analysis starts when CCOD_t events have been observed if CCOD = "event"

CCOD_t

Time difference between analysis start and first patient's enrollment if CCOD = "fixed-first". Time difference between analysis start and last patient's enrollment if CCOD = "fixed-last". Number of events observed when analysis starts if CCOD = "event". Patients enrolled after the analysis start time are excluded from the analysis

etaC

A vector for dropout rate per unit time for control arm

etaE

A vector for dropout rate per unit time for experimental arm. If left NULL, it uses the same dropout rate as eta.

d_itv

A vector of duration of time periods for dropping out the study with rates specified in etaC and etaE. Note that the length of d_itv should be same length as etaC or 1 less.

Value

A .clinClass class containing information on enrollment time, drop-out pattern and analysis start time

Examples

# set the operational parameter values for the trial
# analysis starts at64 time units after first patient in
set_clin(gamma = 10, e_itv = 4, etaC = 0.003,  CCOD = "fixed-first", CCOD_t = 64)
#> Dropout rate per unit time for treatment arm (etaE) is not specified correcly. Disregard this warning if this is for external trial. Otherwise the same dropout rate as eta is used.
#> An object of class ".clinClass"
#> Slot "gamma":
#> [1] 10
#> 
#> Slot "e_itv":
#> [1] 4
#> 
#> Slot "CCOD":
#> [1] "fixed-first"
#> 
#> Slot "CCOD_t":
#> [1] 64
#> 
#> Slot "etaC":
#> [1] 0.003
#> 
#> Slot "etaE":
#> [1] 0.003
#> 
#> Slot "d_itv":
#> NULL
#> 

# analysis starts at 12 time units after last patient in
set_clin(gamma = 2, e_itv = 18, etaC = 0.005,  CCOD = "fixed-last", CCOD_t = 12)
#> Dropout rate per unit time for treatment arm (etaE) is not specified correcly. Disregard this warning if this is for external trial. Otherwise the same dropout rate as eta is used.
#> An object of class ".clinClass"
#> Slot "gamma":
#> [1] 2
#> 
#> Slot "e_itv":
#> [1] 18
#> 
#> Slot "CCOD":
#> [1] "fixed-last"
#> 
#> Slot "CCOD_t":
#> [1] 12
#> 
#> Slot "etaC":
#> [1] 0.005
#> 
#> Slot "etaE":
#> [1] 0.005
#> 
#> Slot "d_itv":
#> NULL
#>