Set up time-to-events

Set up time-to-events

set_event(event, lambdaC, beta, shape, t_itv, change, keep)

Arguments

event

Distribution of time-to-events: event = "pwexp" for piece-wise exponential distribution. event = "weibull" for weibull distribution

lambdaC

Baseline hazard rate of internal control arm. Specify a vector for piece-wise hazard with duration specified in t_itv if event = "pwexp"

beta

covariates' coefficients. NULL if no covariates are found in ssObj

shape

shape of weibull distribution if event = "weibull". NULL if event = "pwexp"

t_itv

a vector indicating interval lengths where the exponential rates provided in lambdaC apply. Note that the length of t_itv is at least 1 less than that of lambdaC and that the final value rate in lambdaC applies after time sum(t_itv). NULL if event = "weibull"

change

operations applied to covariates for simulating time-to-events

keep

original covariates to keep when simulate time-to-events

Value

a .eventClass class containing time-to-events information a matrix containing simulated time-to-events information

Examples

# time-to-event follows a weibull distribution
set_event(event = "weibull", shape = 0.9, lambdaC = 0.0135)
#> An object of class ".eventClass"
#> Slot "event":
#> [1] "weibull"
#> 
#> Slot "lambdaC":
#> [1] 0.0135
#> 
#> Slot "shape":
#> [1] 0.9
#> 
#> Slot "t_itv":
#> NULL
#> 
#> Slot "beta":
#> NULL
#> 
#> Slot "change":
#> NULL
#> 
#> Slot "keep":
#> NULL
#> 

# time-to-event follows a piece-wise exponential distribution
set_event(event = "pwexp", t_itv = 1, lambdaC = c(0.1, 0.02))
#> An object of class ".eventClass"
#> Slot "event":
#> [1] "pwexp"
#> 
#> Slot "lambdaC":
#> [1] 0.10 0.02
#> 
#> Slot "shape":
#> NULL
#> 
#> Slot "t_itv":
#> [1] 1
#> 
#> Slot "beta":
#> NULL
#> 
#> Slot "change":
#> NULL
#> 
#> Slot "keep":
#> NULL
#>